% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrSpline.R
\name{constrSpline}
\alias{constrSpline}
\title{Build a spline constrained function}
\usage{
constrSpline(object)
}
\arguments{
\item{object}{km* model 
constrSpline(object=kmMonotonic1D(design=c(0.1, 0.2, 0.9), response=c(1, 5, 9)))(seq(f=0,t=1,l=100))
constrSpline(object=kmConvex1D(design=c(0.1, 0.5, 0.9), response=c(10, 5, 9)))(seq(f=0,t=1,l=100))}
}
\examples{
design = c(0.05, 0.18, 0.3, 0.5, 0.7, 0.8)
response = c(-5, 2, 4, 5, 9, 15)
object = kmMonotonic1D(design, response)
f = constrSpline(object)
curve(f)
points(design, response, pch=19)
}

