% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmBounded.R
\name{kmBounded1D}
\alias{kmBounded1D}
\title{Kriging model with boundedness constraints}
\usage{
kmBounded1D(design, response, basis.size = dim(design)[1] + 2 + 10,
  covtype = "matern5_2", coef.cov = 0.5 * (max(design) - min(design)),
  coef.var = var(response), lower = min(response) - (max(response) -
  min(response)) * 0.1, upper = max(response) + (max(response) -
  min(response)) * 0.1, nugget = 1e-07 * sd(response))
}
\arguments{
\item{design}{1-column matrix of the design of experiments}

\item{response}{a vector containing the output values given by the real function at the design points}

\item{basis.size}{a value represents the number of the basis functions (descritization of 1D input set)}

\item{covtype}{an optimal character string specifying the covariance function to be used ("gauss" and "matern3_2" choice)}

\item{coef.cov}{a value corresponding to the length theta hyper-parameters of covariance function}

\item{coef.var}{a value specifying the variance parameter}

\item{lower}{lower bound constraint}

\item{upper}{upper bound constraint}

\item{nugget}{an optimal value used as nugget effect to solve the numerical inverse matrix problem}
}
\examples{
model = kmBounded1D(design=c(0.1, 0.3, 0.5, 0.9), response=c(7, -8, 9, 15), lower=-10, upper = 18, coef.cov=1)
}

