% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmConvex1D.R
\name{kmConvex1D}
\alias{kmConvex1D}
\title{Kriging model with convexity 1D constraint}
\usage{
kmConvex1D(design, response, basis.size = dim(design)[1] + 2 + 10,
  covtype = "gauss", coef.cov = 0.5 * (max(design) - min(design)),
  coef.var = var(response), nugget = 1e-07 * sd(response))
}
\arguments{
\item{design}{1-column matrix of the design of experiments}

\item{response}{a vector containing the output values given by the real function at the design points}

\item{basis.size}{a value represents the number of the basis functions (descritization of 1D input set)}

\item{covtype}{an optimal character string specifying the covariance function to be used ("gauss" and "matern3_2" choice)}

\item{coef.cov}{a value corresponding to the length theta hyper-parameters of covariance function}

\item{coef.var}{a value specifying the variance parameter}

\item{nugget}{an optimal value used as nugget effect to solve the numerical inverse matrix problem}
}
\examples{
kmConvex1D(design=c(0.1, 0.5, 0.9), response=c(10, 5, 9), coef.cov = 0.3)
kmConvex1D(design=c(0.1, 0.5,.7, 0.9), response=c(10, 5,7, 9), coef.cov = 0.3)
}

