% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmMonotonic1D.R
\name{kmMonotonic1D}
\alias{kmMonotonic1D}
\title{Kriging model with monotonicity 1D constraint}
\usage{
kmMonotonic1D(design, response, basis.size = dim(design)[1] + 2 + 10,
  covtype = "matern5_2", basis.type = "C1", coef.cov = 0.5 * (max(design)
  - min(design)), coef.var = var(response), nugget = 1e-07 * sd(response))
}
\arguments{
\item{design}{1-column matrix of the design of experiments}

\item{response}{a vector containing the output values given by the real function at the design points}

\item{basis.size}{a value represents the number of the basis functions (descritization of 1D input set)}

\item{covtype}{an optimal character string specifying the covariance function to be used ("gauss" and "matern3_2" choice)}

\item{basis.type}{an optimal character string specifying the regularization of the basis functions to be used (which implies the class of the simulation paths)}

\item{coef.cov}{a value corresponding to the length theta hyper-parameters of covariance function}

\item{coef.var}{a value specifying the variance parameter}

\item{nugget}{an optimal value used as nugget effect to solve the numerical inverse matrix problem}
}
\examples{
kmMonotonic1D(design=c(0.1, 0.5, 0.9), response=c(1, 5, 9))
kmMonotonic1D(design=c(0.1, 0.5, 0.9), response=c(1, 5, 9), basis.type="C0")
kmMonotonic1D(design=sort(runif(5)), response=cumsum(runif(5)), nugget=1e-4, covtype="gauss")
f <- function(x){
log(20*x+1)
}
design <- c(0, 0.1, 0.2, 0.3, 0.4, 0.9, 1)
response <- f(design)
meany <- mean(response)
f <- function(x){
log(20*x+1)-meany
}
design <- c(0, 0.1, 0.2, 0.3, 0.4, 0.9, 1)
response <- f(design)
model = kmMonotonic1D(design, response, coef.var=335^2, coef.cov=4.7,basis.size=50)
}

