% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kmBounded.R
\name{Phi1D.kmBounded}
\alias{Phi1D.kmBounded}
\title{Apply basis functions to given data in design space}
\usage{
\method{Phi1D}{kmBounded}(model, newdata)
}
\arguments{
\item{model}{km* model}

\item{newdata}{data in design space}
}
\description{
Apply basis functions to given data in design space
}

