% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmBounded.R
\name{Phi1D.kmBounded1D}
\alias{Phi1D.kmBounded1D}
\title{Apply basis functions to given data in design space}
\usage{
\method{Phi1D}{kmBounded}(model, newdata)
}
\arguments{
\item{model}{km* model}

\item{newdata}{data in design space}
}

