% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmMonotonic1D.R
\name{Phi1D.kmMonotonic1D}
\alias{Phi1D.kmMonotonic1D}
\title{Apply basis functions to given data in design space}
\usage{
\method{Phi1D}{kmMonotonic1D}(model, newdata)
}
\arguments{
\item{model}{km* model}

\item{newdata}{data in design space}
}

