% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/constrSpline.R
\name{Phi1D}
\alias{Phi1D}
\title{Generic Phi1D.km* function}
\usage{
Phi1D(model, newdata)
}
\arguments{
\item{model}{km* model}

\item{newdata}{data in design space}
}
\description{
Generic Phi1D.km* function
}

