% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot the km* object}
\usage{
plot(x = seq(f = min(object$call$design), t = max(object$call$design), l =
  100), object, spline = TRUE, nsim = 100, median = TRUE, mean = FALSE,
  minmax = FALSE, quantiles = TRUE, col = "black", add = F)
}
\arguments{
\item{x}{The coordinates to plot in the object design space (seq(0,1,,100) by default)}

\item{object}{The km* model to plot}

\item{spline}{Whether to plot or not the spline of the model (by default TRUE)}

\item{nsim}{Number of simulations of the model realizations to build the sample in order to get the median, mean, minmax and quantiles statistics (by default 100)}

\item{median}{Whether to plot or not the median of the model sample (by default TRUE)}

\item{mean}{Whether to plot or not the mean of the model sample (by default FALSE)}

\item{minmax}{Whether to plot or not the min/max of the model sample (by default FALSE)}

\item{quantiles}{Whether to plot or not the 1st and 3rd quantiles of the model sample (by default TRUE)}

\item{col}{Color to use for drawing}

\item{add}{Should we add this plot to an existing one ?}
}
\description{
Plot the km* object
}
\examples{
design = c(0.1, 0.5, 0.9)
response = c(10, 5, 9)
model = kmConvex1D(design, response)
plot(object=model, median=TRUE, quantiles=TRUE, minmax=FALSE, col='red')
}

