% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictCK}
\alias{predictCK}
\title{Predict response from a km* model}
\usage{
predictCK(object, newdata, nsim = 100)
}
\arguments{
\item{object}{km* model}

\item{newdata}{a vector which represents the points where to performs predictions}

\item{nsim}{the number of response vectors to simulate}
}
\examples{
predict(object=kmMonotonic1D(design=c(0.1, 0.5, 0.9), response=c(1, 5, 9)), newdata=seq(f=0, t=1, l=100), nsim=10)
predict(object=kmConvex1D(design=c(0.1, 0.5, 0.9), response=c(5, 1, 9)), newdata=seq(f=0, t=1, l=100), nsim=10)
design = c(0.1, 0.5, 0.9)
response = c(1, 8, 9)
model = kmMonotonic1D(design, response)
graphics::plot(x=seq(0,1,,100),y=predict(object=model, newdata=seq(0,1,,100), nsim=100)[,'Median'], type='l')
points(design, response, pch=19)
}

