% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate.kmBounded.R
\name{simulate_process.kmBounded}
\alias{simulate_process.kmBounded}
\title{Simulate responses vectors from a kmBounded model}
\usage{
\method{simulate_process}{kmBounded}(object, nsim, seed = NULL, newdata)
}
\arguments{
\item{object}{kmBounded model}

\item{nsim}{the number of response vectors to simulate}

\item{seed}{optional random seed}

\item{newdata}{a vector which represents the points where to performs predictions}
}
\description{
Simulate responses vectors from a kmBounded model
}
\examples{
design=c(0.1, 0.3, 0.5, 0.9)
response = c(7, -8, 9, 15)
model = kmBounded(design, response, lower = -10, upper = 18)
x=seq(0,1,,100)
graphics::plot(x=x,y=simulate_process(object=model, newdata=x, nsim=1),type='l')
}

