% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate.kmConvex1D.R
\name{simulate_process.kmConvex1D}
\alias{simulate_process.kmConvex1D}
\title{Simulate responses vectors from a kmConvex1D object}
\usage{
\method{simulate_process}{kmConvex1D}(object, nsim, seed = NULL, newdata)
}
\arguments{
\item{object}{kmConvex1D model}

\item{nsim}{the number of response vectors to simulate}

\item{seed}{optional random seed}

\item{newdata}{a vector which represents the points where to performs predictions}
}
\description{
Simulate responses vectors from a kmConvex1D object
}
\examples{
design = c(0.1, 0.5, 0.9)
response = c(8, 5, 9)
model = kmConvex1D(design, response)
graphics::plot(x=seq(0,1,,100),y=simulate_process(object=model, newdata=seq(0,1,,100), nsim=1))
}

