% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{simulate_process}
\alias{simulate_process}
\title{Generic simulate_process method}
\usage{
simulate_process(object, nsim, seed, newdata)
}
\arguments{
\item{object}{km* model}

\item{nsim}{number of simulated processes}

\item{seed}{random seed}

\item{newdata}{data in design space where to evaluate simlations}
}

