% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/constrSpline.R
\name{constrSpline}
\alias{constrSpline}
\title{Build a spline constrained function}
\usage{
constrSpline(object)
}
\arguments{
\item{object}{km* model
constrSpline(object=kmMonotonic1D(design=c(0.1, 0.5, 0.9), response=c(1, 5, 9)))(seq(f=0,t=1,l=100))
constrSpline(object=kmConvex1D(design=c(0.1, 0.5, 0.9), response=c(1, 5, 9)))(seq(f=0,t=1,l=100))}
}
\description{
Build a spline constrained function
}
\examples{
design = c(0.1, 0.5, 0.9)
response = c(1, 5, 9)
object = kmMonotonic1D(design, response, basis.type="C1")
f = constrSpline(object)
curve(f)
}

