% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kmMonotonic1D.R
\name{kmMonotonic1D}
\alias{kmMonotonic1D}
\title{Kriging model with monotonicity 1D constraint}
\usage{
kmMonotonic1D(design, response, basis.size = dim(design)[1] + 2 + 10,
  covtype = "gauss", basis.type = "C0", coef.cov = 0.5 * (max(design) -
  min(design)), coef.var = var(response), nugget = 1e-08)
}
\arguments{
\item{design}{1-column matrix of the design of experiments}

\item{response}{a vector containing the output values given by the real function at the design points}

\item{basis.size}{a value represents the number of the basis functions (descritization of 1D input set)}

\item{covtype}{an optimal character string specifying the covariance function to be used ("gauss" and "matern3_2" choice)}

\item{basis.type}{an optimal character string specifying the regularization of the basis functions to be used (which implies the class of the simulation paths)}

\item{coef.cov}{a value corresponding to the length theta hyper-parameters of covariance function}

\item{coef.var}{a value specifying the variance parameter}

\item{nugget}{an optimal value used as nugget effect to solve the numerical inverse matrix problem}
}
\description{
Kriging model with monotonicity 1D constraint
}

