% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.kmConvex1D.R
\name{simulate_process.kmConvex1D}
\alias{simulate_process.kmConvex1D}
\title{Simulate responses vectors from a kmConvex1D object}
\usage{
\method{simulate_process}{kmConvex1D}(object, nsim, seed = NULL, newdata)
}
\arguments{
\item{object}{kmConvex1D model}

\item{nsim}{the number of response vectors to simulate}

\item{seed}{optional random seed}

\item{newdata}{a vector which represents the points where to performs predictions}
}
\examples{
design = c(0.1, 0.5, 0.9)
response = c(10, 5, 9)
model = kmConvex1D(design, response, coef.cov = 0.35)
x = seq(0, 1,, 100)
graphics::matplot(x, y=simulate_process(object=model, newdata=x, nsim=100), type='l', col='gray', lty=1, ylab='response')
lines(x,constrSpline(object=model)(x), lty=1, col='black')
points(design, response, pch=19)
legend(0.15, 12.5, c("convex GP sample paths", "posterior max"), 
       col = c('gray', 'black'), text.col = "black",
       lty = c(1, 1), pch=c(NA_integer_, NA_integer_), lwd = c(2, 2), text.font=1, box.lty=0, cex=1)
}

